Change file name and rename createLibP2p
