+++ packages/lodestar/test/unit/cli/commands/dev.test.ts
